+++ Object/Brdf/Lambert.cpp
@@ -22,7 +22,7 @@ namespace Object {
 
         Object::Color Lambert::reflected(const Math::Vector &, const Math::Normal &, const Math::Vector &, const Object::Color &albedo) const
         {
-            return albedo * mStrength / M_PI;
+            return albedo * mStrength / (float)M_PI;
         }
 
         Math::Vector Lambert::sample(Render::Sampler &sampler, const Math::Normal &normal, const Math::Vector &) const
